refactor: remove components, service and interfaces  into src
