+++ rs-react-app/src/components/result-block/ResultBlock.tsx
