+++ rs-react-app/src/components/search-block/SearchBlock.css
