+++ rs-react-app/src/interfaces/types.tsx
